バックエンドリファクタリング: Entityを経由するよう変更

+++ packages/backend/src/reminder/reminder.service.ts
@@ -4,6 +4,7 @@ import { Reminder } from '@prisma/client';
 
 import { PrismaService } from '../vendors/prisma/prisma.service';
 import { ReminderCreationDto, ReminderResponseDto } from './reminder.dto';
+import { ReminderEntity } from './reminder.entity';
 
 @Injectable()
 export class ReminderService {
@@ -13,45 +14,41 @@ export class ReminderService {
     return 'Hello World!';
   }
 
-  async findAll(): Promise<Reminder[]> {
-    return await this.prisma.reminder.findMany({
+  async findAll(): Promise<ReminderEntity[]> {
+    const models = await this.prisma.reminder.findMany({
       where: {
-        completedAt: null
-      }
-    });
-  }
-
-  async findByUuid(uuid: string): Promise<Reminder> {
-    return await this.prisma.reminder.findUnique({ where: { uuid } });
-  }
-
-  async create(creationDto: ReminderCreationDto): Promise<Reminder> {
-    return await this.prisma.reminder.create({
-      data: {
-        ...creationDto,
-        uuid: v4(),
-        createdAt: new Date(),
+        completedAt: null,
       },
     });
-  }
-
-  async remove(reminder: Reminder): Promise<void> {
-    await this.prisma.reminder.delete({ where: { uuid: reminder.uuid } });
-  }
-
-  async complete(reminder: Reminder): Promise<void> {
-    await this.prisma.reminder.update({
-      where: { uuid: reminder.uuid },
-      data: { completedAt: new Date() },
+    return models.map((model) => {
+      return ReminderEntity.fromPrisma(model);
     });
   }
 
-  toResponse(reminder: Reminder): ReminderResponseDto {
-    return {
-      uuid: reminder.uuid,
-      title: reminder.title,
-      description: reminder.description,
-      datetime: reminder.datetime.toISOString(),
-    };
+  async findByUuid(uuid: string): Promise<ReminderEntity> {
+    const reminder = await this.prisma.reminder.findUnique({ where: { uuid } });
+    if (reminder === null) {
+    }
+    return ReminderEntity.fromPrisma(reminder);
+  }
+
+  async create(creationDto: ReminderCreationDto): Promise<ReminderEntity> {
+    const reminder = ReminderEntity.factory(creationDto);
+    return await this.save(reminder);
+  }
+
+  async save(reminder: ReminderEntity): Promise<ReminderEntity> {
+    if (reminder.isRemoved) {
+      await this.prisma.reminder.delete({ where: { uuid: reminder.uuid } });
+      return;
+    }
+    const model = await this.prisma.reminder.upsert({
+      where: {
+        uuid: reminder.uuid,
+      },
+      update: reminder.toPrisma(),
+      create: reminder.toPrisma(),
+    });
+    return ReminderEntity.fromPrisma(model);
   }
 }
